+++ CHANGELOG.md
@@ -4,6 +4,20 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.1.0] - 2020-03-28
+
+### Added
+
+- Handling a click on the track to move the thumb.
+
+### Changed
+
+- The thumbs can no longer cross each other. 
+
+### Fixed
+
+- Issue [#15](https://github.com/AkhmadBabaev/ordinary-slider/issues/15).
+
 ## [3.0.1] - 2020-03-27
 
 ### Fixed
@@ -21,7 +35,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Removed 
 
-- Method reset
+- Method reset.
 
 ### Fixed
 
@@ -56,6 +70,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 First working release.
 
+[3.1.0]: https://github.com/AkhmadBabaev/ordinary-slider/compare/3.0.1...3.1.0
 [3.0.1]: https://github.com/AkhmadBabaev/ordinary-slider/compare/3.0.0...3.0.1
 [3.0.0]: https://github.com/AkhmadBabaev/ordinary-slider/compare/2.0.0...3.0.0
 [2.0.0]: https://github.com/AkhmadBabaev/ordinary-slider/compare/1.2.0...2.0.0